Bug 463293 - Babel should use OpenID
Provide a new login method using Eclipse OpenID. To keep the number of
changes low the users table is still used. When a user logged in through
OpenID we request some user details from account API to insert/update
the user in database. (but only parts of the existing user table)

Change-Id: Ifad61fe7fc3b3df3c1a89258b83da6735607fb7e
Signed-off-by: Paul Pazderski <paul-eclipse@ppazderski.de>
Signed-off-by: droy <denis.roy@eclipse-foundation.org>